Extract UploadResource class
Removes duplication from Process and ModelProcess
