+++ lib/sdr_client/deposit.rb
@@ -59,6 +59,7 @@ require 'sdr_client/deposit/file'
 require 'sdr_client/deposit/file_metadata_builder'
 require 'sdr_client/deposit/file_set'
 require 'sdr_client/deposit/request'
-require 'sdr_client/deposit/upload_files'
 require 'sdr_client/deposit/metadata_builder'
 require 'sdr_client/deposit/process'
+require 'sdr_client/deposit/upload_files'
+require 'sdr_client/deposit/upload_resource'
