+++ lib/sdr_client/deposit/model_process.rb
@@ -6,7 +6,6 @@ module SdrClient
   module Deposit
     # The process for doing a deposit from a Cocina Model
     class ModelProcess
-      DRO_PATH = '/v1/resources'
       # @param [Cocina::Model::RequestDRO] request_dro for depositing
       # @param [Connection] connection the connection to use
       # @param [Array<String>] files a list of file names to upload
@@ -30,16 +29,15 @@ module SdrClient
                                            connection: connection,
                                            mime_types: mime_types).run
         new_request_dro = with_external_identifiers(upload_responses)
-        upload_request_dro(new_request_dro.to_json)
+        UploadResource.run(accession: @accession,
+                           metadata: new_request_dro.to_json,
+                           logger: logger,
+                           connection: connection)
       end
 
       private
 
       attr_reader :request_dro, :files, :logger, :connection
-
-      def accession?
-        @accession
-      end
 
       def check_files_exist
         logger.info('checking to see if files exist')
@@ -61,30 +59,6 @@ module SdrClient
         request_files.keys.each do |request_filename|
           raise "File not provided for request file #{request_filename}" unless filenames.include?(request_filename)
         end
-      end
-
-      # @return [Hash<Symbol,String>] the result of the metadata call
-      def upload_request_dro(request_json)
-        logger.info("Starting upload metadata: #{request_json}")
-        response = connection.post(path, request_json, 'Content-Type' => 'application/json')
-        unexpected_response(response) unless response.status == 201
-
-        logger.info("Response from server: #{response.body}")
-
-        { druid: JSON.parse(response.body)['druid'], background_job: response.headers['Location'] }
-      end
-
-      def path
-        path = DRO_PATH
-        path += '?accession=true' if accession?
-        path
-      end
-
-      def unexpected_response(response)
-        raise "There was an error with your request: #{response.body}" if response.status == 400
-        raise 'There was an error with your credentials. Perhaps they have expired?' if response.status == 401
-
-        raise "unexpected response: #{response.status} #{response.body}"
       end
 
       # Map of filenames to mimetypes
